+++ roseau/load_flow/models/sources.py
@@ -6,7 +6,7 @@ from typing import Final
 import numpy as np
 from typing_extensions import Self
 
-from roseau.load_flow.converters import _PHASE_SIZES, _calculate_voltages, calculate_voltage_phases
+from roseau.load_flow.converters import _PHASE_SIZES, ALPHA, _calculate_voltages, calculate_voltage_phases
 from roseau.load_flow.exceptions import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
 from roseau.load_flow.models.buses import Bus
 from roseau.load_flow.models.core import Element
@@ -18,7 +18,14 @@ logger = logging.getLogger(__name__)
 
 
 class VoltageSource(Element):
-    """A voltage source."""
+    """A voltage source fixes the voltages on the phases of the bus it is connected to.
+
+    The source can be connected in a wye or star configuration (i.e with a neutral) or in a delta
+    configuration (i.e without a neutral).
+
+    See Also:
+        The :ref:`Voltage source documentation page <models-voltage-source-usage>` for example usage.
+    """
 
     allowed_phases: Final = Bus.allowed_phases
     """The allowed phases for a voltage source are the same as for a :attr:`bus<Bus.allowed_phases>`."""
@@ -28,7 +35,7 @@ class VoltageSource(Element):
         id: Id,
         bus: Bus,
         *,
-        voltages: ComplexArrayLike1D,
+        voltages: ComplexArrayLike1D | complex | Q_[complex],
         phases: str | None = None,
         connect_neutral: bool | None = None,
     ) -> None:
@@ -42,10 +49,19 @@ class VoltageSource(Element):
                 The bus of the voltage source.
 
             voltages:
-                An array-like of the voltages of the source. They will be set on the connected bus.
-                If the source has a neutral connection, the voltages are considered phase-to-neutral
-                voltages, otherwise they are the phase-to-phase voltages. Either complex values (V)
-                or a :class:`Quantity <roseau.load_flow.units.Q_>` of complex values.
+                A single voltage value or an array-like of the voltages of the source to be fixed on
+                the connected bus phases. If the source has a neutral connection, the voltages are
+                considered phase-to-neutral voltages, otherwise they are the phase-to-phase voltages.
+                Either pass complex values (V) or a :class:`Quantity <roseau.load_flow.units.Q_>` of
+                complex values.
+
+                When a scalar value is passed, it is interpreted as the first value of the source
+                voltages vector. The other values are calculated based on the number of phases of
+                the source. For a single-phase source, the passed scalar value is used. For a two-
+                phase source, the second voltage value is the negative of the first value (180°
+                phase shift). For a three-phase source, the second and third values are calculated
+                based on the first value and the phase shift of -120° and 120°, respectively (120°
+                phase shift clockwise).
 
             phases:
                 The phases of the source. A string like ``"abc"`` or ``"an"`` etc. The bus phases are
@@ -121,17 +137,40 @@ class VoltageSource(Element):
     @property
     @ureg_wraps("V", (None,))
     def voltages(self) -> Q_[ComplexArray]:
-        """The voltages of the source (V)."""
+        """The complex voltages of the source (V).
+
+        Setting the voltages will update the source voltages and invalidate the network results.
+
+        Note:
+            Setting a scalar value updates the complex voltages of all phases of the source, not
+            just their magnitudes. The phase angles are calculated based on the number of phases of
+            the source. For a single-phase source, the phase angle is 0°. For a two-phase source,
+            the phase angle of the second phase is 180°. For a three-phase source, the phase angles
+            of the second and third phases are -120° and 120°, respectively (120° phase shift
+            clockwise).
+        """
         return self._voltages
 
     @voltages.setter
     @ureg_wraps(None, (None, "V"))
-    def voltages(self, voltages: ComplexArrayLike1D) -> None:
+    def voltages(self, value: ComplexArrayLike1D | complex | Q_[complex]) -> None:
+        """Set the voltages of the source."""
+        if np.isscalar(value):
+            if self._size == 1:
+                voltages = [value]
+            elif self._size == 2:
+                voltages = [value, -value]
+            else:
+                assert self._size == 3
+                voltages = [value, value * ALPHA**2, value * ALPHA]
+        else:
+            voltages = value
+        voltages = np.array(voltages, dtype=np.complex128)
         if len(voltages) != self._size:
             msg = f"Incorrect number of voltages: {len(voltages)} instead of {self._size}"
             logger.error(msg)
             raise RoseauLoadFlowException(msg, code=RoseauLoadFlowExceptionCode.BAD_VOLTAGES_SIZE)
-        self._voltages = np.array(voltages, dtype=np.complex128)
+        self._voltages = voltages
         self._invalidate_network_results()
         if self._cy_element is not None:
             self._cy_element.update_voltages(self._voltages)